+++ src/FECFileLoader.py
@@ -12,17 +12,168 @@ import fecfile
 import json
 import os
 import requests
+from collections import OrderedDict
 from typing import Any, Dict, List
-from src import JSONType, logger, schema
+from psycopg2 import sql
+from psycopg2.sql import SQL, Literal
+from src import JSONType, logger
 from src.database import Database
 from src.sqs import delete_message_from_sqs, parse_message
 
 
-# business logic
 FILING_TYPE = os.environ['FILING_TYPE']
 
-def upsert_schedule_b_filing(fec_file_id: str, filing: Dict[str, Any]) -> bool:
-    """upserts a single schedule B filing
+#
+# schedule B schema
+#
+
+def schedule_b_exists(transaction_id_number: str) -> SQL:
+    """returns a query to check if a transaction id has a record
+
+    Args:
+        transaction_id_number (str): ID representing transaction
+
+    Returns:
+        SQL: select query for record
+    """
+
+    query = sql.SQL('SELECT * FROM fec.filings_schedule_b WHERE transaction_id_number={}')\
+                .format(Literal(transaction_id_number))
+
+    return query
+
+
+def schedule_b_insert(fec_file_id: str, filing: Dict[str, Any]) -> SQL:
+    """inserts a record into fec.filings_schedule_b
+
+    Args:
+        fec_file_id (str): filing ID
+        filing (Dict[str, Any]): dictionary containing transaction data of filing
+
+    Returns:
+        SQL: SQL insert query
+    """
+
+    filing['fec_file_id'] = fec_file_id
+    values = OrderedDict(sorted(filing.items()))
+
+    query_string = 'INSERT INTO fec.filings_schedule_b ('\
+        + ', '.join([f'{key}' for key, val in values.items()])\
+        + ') '\
+        + 'VALUES ('\
+        + ', '.join(['{}' for key, val in values.items()])\
+        + ')'
+
+    query = sql.SQL(query_string)\
+                .format(*[Literal(val) for key, val in values.items()])
+
+    return query
+
+#
+# schedule E schema
+#
+
+def schedule_e_exists(transaction_id_number: str) -> SQL:
+    """returns a query to check if a transaction id has a record
+
+    Args:
+        transaction_id_number (str): ID representing transaction
+
+    Returns:
+        SQL: select query for record
+    """
+
+    query = sql.SQL('SELECT * FROM fec.filings_schedule_e WHERE transaction_id_number={}')\
+                .format(Literal(transaction_id_number))
+
+    return query
+
+
+def schedule_e_insert(fec_file_id: str, filing: Dict[str, Any]) -> SQL:
+    """inserts a record into fec.filings_schedule_e
+
+    Args:
+        fec_file_id (str): filing ID
+        filing (Dict[str, Any]): dictionary containing transaction data of filing
+
+    Returns:
+        SQL: SQL insert query
+    """
+
+    filing['fec_file_id'] = fec_file_id
+    values = OrderedDict(sorted(filing.items()))
+
+    query_string = 'INSERT INTO fec.filings_schedule_e ('\
+        + ', '.join([f'{key}' for key, val in values.items()])\
+        + ') '\
+        + 'VALUES ('\
+        + ', '.join(['{}' for key, val in values.items()])\
+        + ')'
+
+    query = sql.SQL(query_string)\
+                .format(*[Literal(val) for key, val in values.items()])
+
+    return query
+
+#
+# Form 1 Supplemental schema
+#
+
+def f1_supplemental_exists(fec_file_id: str, filing: Dict[str, Any]) -> SQL:
+    """checks for existining supplemental data
+
+    Args:
+        fec_file_id (str): Filing ID
+        filing (Dict[str, Any]): dictionary containing transaction data of filing
+
+    Returns:
+        SQL: select query
+    """
+
+    filing['fec_file_id'] = fec_file_id
+    values = OrderedDict(sorted(filing.items()))
+
+    query_string = 'SELECT * FROM fec.form_1_supplemental WHERE ' \
+        + ' AND '.join([f' {key}={{}}' for key, val in values.items()])
+
+    query = sql.SQL(query_string)\
+        .format(*[Literal(val) for key, val in values.items()])
+
+    return query
+
+
+def f1_supplemental_insert(fec_file_id: str, filing: Dict[str, Any]) -> SQL:
+    """inserts a record into fec.form_1_supplemental
+
+    Args:
+        fec_file_id (str): filing ID
+        filing (Dict[str, Any]): dictionary containing transaction data of filing
+
+    Returns:
+        SQL: SQL insert query
+    """
+
+    filing['fec_file_id'] = fec_file_id
+    values = OrderedDict(sorted(filing.items()))
+
+    query_string = 'INSERT INTO fec.form_1_supplemental ('\
+        + ', '.join([f'{key}' for key, val in values.items()])\
+        + ') '\
+        + 'VALUES ('\
+        + ', '.join(['{}' for key, val in values.items()])\
+        + ')'
+
+    query = sql.SQL(query_string)\
+                .format(*[Literal(val) for key, val in values.items()])
+
+    return query
+
+#
+# business logic
+#
+
+def insert_schedule_b_filing(fec_file_id: str, filing: Dict[str, Any]) -> bool:
+    """inserts a single schedule B filing
 
     Args:
         fec_file_id (str): FEC filing ID
@@ -33,7 +184,7 @@ def upsert_schedule_b_filing(fec_file_id: str, filing: Dict[str, Any]) -> bool:
     """
 
     pk = filing['transaction_id_number']
-    exists_query = schema.schedule_b_exists(pk)
+    exists_query = schedule_b_exists(pk)
     with Database() as db:
         record_exists = db.record_exists(exists_query)
         if record_exists:
@@ -42,13 +193,13 @@ def upsert_schedule_b_filing(fec_file_id: str, filing: Dict[str, Any]) -> bool:
             return True
 
         else:
-            query = schema.schedule_b_insert(fec_file_id, filing)
+            query = schedule_b_insert(fec_file_id, filing)
 
             return db.try_query(query)
 
 
-def upsert_schedule_e_filing(fec_file_id: str, filing: Dict[str, Any]) -> bool:
-    """upserts a single schedule E filing
+def insert_schedule_e_filing(fec_file_id: str, filing: Dict[str, Any]) -> bool:
+    """inserts a single schedule E filing
 
     Args:
         fec_file_id (str): FEC filing ID
@@ -59,7 +210,7 @@ def upsert_schedule_e_filing(fec_file_id: str, filing: Dict[str, Any]) -> bool:
     """
 
     pk = filing['transaction_id_number']
-    exists_query = schema.schedule_e_exists(pk)
+    exists_query = schedule_e_exists(pk)
     with Database() as db:
         record_exists = db.record_exists(exists_query)
         if record_exists:
@@ -68,13 +219,13 @@ def upsert_schedule_e_filing(fec_file_id: str, filing: Dict[str, Any]) -> bool:
             return True
 
         else:
-            query = schema.schedule_e_insert(fec_file_id, filing)
+            query = schedule_e_insert(fec_file_id, filing)
 
             return db.try_query(query)
 
 
-def upsert_f1_supplemental(fec_file_id: str, filing: Dict[str, Any]) -> bool:
-    """upserts a single filing of Form 1 Supplemental Data
+def insert_f1_supplemental(fec_file_id: str, filing: Dict[str, Any]) -> bool:
+    """inserts a single filing of Form 1 Supplemental Data
 
     Args:
         fec_file_id (str): FEC filing ID
@@ -84,7 +235,7 @@ def upsert_f1_supplemental(fec_file_id: str, filing: Dict[str, Any]) -> bool:
         bool: if in database
     """
 
-    exists_query = schema.f1_supplemental_exists(fec_file_id, filing)
+    exists_query = f1_supplemental_exists(fec_file_id, filing)
     with Database() as db:
         record_exists = db.record_exists(exists_query)
         if record_exists:
@@ -92,14 +243,14 @@ def upsert_f1_supplemental(fec_file_id: str, filing: Dict[str, Any]) -> bool:
             return True
 
         else:
-            query = schema.f1_supplemental_insert(fec_file_id, filing)
+            query = f1_supplemental_insert(fec_file_id, filing)
 
             return db.try_query(query)
 
 
 
-def upsert_filing(fec_file_id: str, filing: Dict[str, Any]) -> bool:
-    """upserts a single filing
+def insert_filing(fec_file_id: str, filing: Dict[str, Any]) -> bool:
+    """inserts a single filing
 
     Args:
         fec_file_id (str): FEC filing ID
@@ -114,18 +265,18 @@ def upsert_filing(fec_file_id: str, filing: Dict[str, Any]) -> bool:
     # Schedule B Filings
     if form_type.startswith('SB'):
 
-        return upsert_schedule_b_filing(fec_file_id, filing)
+        return insert_schedule_b_filing(fec_file_id, filing)
 
 
     # Schedule E Filings
     elif form_type.startswith('SE'):
 
-        return upsert_schedule_e_filing(fec_file_id, filing)
+        return insert_schedule_e_filing(fec_file_id, filing)
 
     # Form 1 Supplemental Data Filings
     elif form_type.startswith('F1S'):
 
-        return upsert_f1_supplemental(fec_file_id, filing)
+        return insert_f1_supplemental(fec_file_id, filing)
 
     else:
         logger.error(f'Filing of form_type {form_type} does not match those available')
@@ -158,7 +309,7 @@ def lambdaHandler(event:dict, context: object) -> bool:
                                 options={'filter_itemizations': [FILING_TYPE]}):
 
             if fec_item.data_type == 'itemization':
-                upsert_filing(filing_id, fec_item.data)
+                insert_filing(filing_id, fec_item.data)
 
         delete_message_from_sqs(message)
