+++ src/schema.py
@@ -383,52 +383,6 @@ def update_committee_total(committee_total: JSONType) -> SQL:
     return query
 
 #
-# schedule B filings
-#
-
-def schedule_b_exists(transaction_id_number: str) -> SQL:
-    """returns a query to check if a transaction id has a record
-
-    Args:
-        transaction_id_number (str): ID representing transaction
-
-    Returns:
-        SQL: select query for record
-    """
-
-    query = sql.SQL('SELECT * FROM fec.filings_schedule_b WHERE transaction_id_number={}')\
-                .format(Literal(transaction_id_number))
-
-    return query
-
-
-def schedule_b_insert(fec_file_id: str, filing: Dict[str, Any]) -> SQL:
-    """inserts a record into fec.filings_schedule_b
-
-    Args:
-        fec_file_id (str): filing ID
-        filing (Dict[str, Any]): dictionary containing transaction data of filing
-
-    Returns:
-        SQL: SQL insert query
-    """
-
-    filing['fec_file_id'] = fec_file_id
-    values = OrderedDict(sorted(filing.items()))
-
-    query_string = 'INSERT INTO fec.filings_schedule_b ('\
-        + ', '.join([f'{key}' for key, val in values.items()])\
-        + ') '\
-        + 'VALUES ('\
-        + ', '.join(['{}' for key, val in values.items()])\
-        + ')'
-
-    query = sql.SQL(query_string)\
-                .format(*[Literal(val) for key, val in values.items()])
-
-    return query
-
-#
 # Candidates
 #
 
@@ -492,106 +446,5 @@ def candidate_update(candidate: Dict[str, str]) -> SQL:
 
     query = sql.SQL(query_string)\
         .format(*[Literal(val) for key, val in values.items()], Literal(primary_key))
-
-    return query
-
-
-#
-# schedule E filings
-#
-
-def schedule_e_exists(transaction_id_number: str) -> SQL:
-    """returns a query to check if a transaction id has a record
-
-    Args:
-        transaction_id_number (str): ID representing transaction
-
-    Returns:
-        SQL: select query for record
-    """
-
-    query = sql.SQL('SELECT * FROM fec.filings_schedule_e WHERE transaction_id_number={}')\
-                .format(Literal(transaction_id_number))
-
-    return query
-
-
-def schedule_e_insert(fec_file_id: str, filing: Dict[str, Any]) -> SQL:
-    """inserts a record into fec.filings_schedule_e
-
-    Args:
-        fec_file_id (str): filing ID
-        filing (Dict[str, Any]): dictionary containing transaction data of filing
-
-    Returns:
-        SQL: SQL insert query
-    """
-
-    filing['fec_file_id'] = fec_file_id
-    values = OrderedDict(sorted(filing.items()))
-
-    query_string = 'INSERT INTO fec.filings_schedule_e ('\
-        + ', '.join([f'{key}' for key, val in values.items()])\
-        + ') '\
-        + 'VALUES ('\
-        + ', '.join(['{}' for key, val in values.items()])\
-        + ')'
-
-    query = sql.SQL(query_string)\
-                .format(*[Literal(val) for key, val in values.items()])
-
-    return query
-
-
-#
-# Form 1 Supplemental Data
-#
-
-def f1_supplemental_exists(fec_file_id: str, filing: Dict[str, Any]) -> SQL:
-    """checks for existining supplemental data
-
-    Args:
-        fec_file_id (str): Filing ID
-        filing (Dict[str, Any]): dictionary containing transaction data of filing
-
-    Returns:
-        SQL: select query
-    """
-
-    filing['fec_file_id'] = fec_file_id
-    values = OrderedDict(sorted(filing.items()))
-
-    query_string = 'SELECT * FROM fec.form_1_supplemental WHERE ' \
-        + ' AND '.join([f' {key}={{}}' for key, val in values.items()])
-
-    query = sql.SQL(query_string)\
-        .format(*[Literal(val) for key, val in values.items()])
-
-    return query
-
-
-def f1_supplemental_insert(fec_file_id: str, filing: Dict[str, Any]) -> SQL:
-    """inserts a record into fec.form_1_supplemental
-
-    Args:
-        fec_file_id (str): filing ID
-        filing (Dict[str, Any]): dictionary containing transaction data of filing
-
-    Returns:
-        SQL: SQL insert query
-    """
-
-    filing['fec_file_id'] = fec_file_id
-    values = OrderedDict(sorted(filing.items()))
-
-    query_string = 'INSERT INTO fec.form_1_supplemental ('\
-        + ', '.join([f'{key}' for key, val in values.items()])\
-        + ') '\
-        + 'VALUES ('\
-        + ', '.join(['{}' for key, val in values.items()])\
-        + ')'
-
-    query = sql.SQL(query_string)\
-                .format(*[Literal(val) for key, val in values.items()])
 
     return query
